+++ springboot/test/src/test/java/org/kie/kogito/test/springboot/kafka/KafkaTestClientTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito.test.kafka;
+package org.kie.kogito.test.springboot.kafka;
 
 import java.time.Duration;
 import java.util.ArrayList;
